[Fix] - 저장된 답변을 불러올때 500 에러가 뜨는 문제

+++ src/main/java/com/heartz/byeboo/domain/type/EStep.java
@@ -6,17 +6,17 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 @Getter
 public enum EStep {
-    EXPRESS_EMOTION("감정 쏟아내기", 1),
-    ORGANIZE_SITUATION("상황 정리하기", 2),
-    REFLECT_ROLE("내 역할 돌아보기", 3),
-    FIND_MEANING("관계의 의미 찾기", 4),
-    UNDERSTAND_SELF("지금의 나 이해하기", 5),
-    AWAKE_HEART("마음 깨우기", 1),
-    LET_FLOW_EMOTION("감정 흘려보내기", 2),
-    NAMING_EMOTION("감정에 이름 붙이기", 3),
-    SMALL_ROUTINE("나를 위한 작은 루틴", 4),
+    EXPRESS_EMOTION("감정 쏟아내기", 1L),
+    ORGANIZE_SITUATION("상황 정리하기", 2L),
+    REFLECT_ROLE("내 역할 돌아보기", 3L),
+    FIND_MEANING("관계의 의미 찾기", 4L),
+    UNDERSTAND_SELF("지금의 나 이해하기", 5L),
+    AWAKE_HEART("마음 깨우기", 1L),
+    LET_FLOW_EMOTION("감정 흘려보내기", 2L),
+    NAMING_EMOTION("감정에 이름 붙이기", 3L),
+    SMALL_ROUTINE("나를 위한 작은 루틴", 4L),
     ;
 
     private final String label;
-    private final Integer stepNumber;
+    private final Long stepNumber;
 }
